Error from protector in PHP 7.1

+++ htdocs/xoops_lib/modules/protector/include/precheck.inc.php
@@ -9,7 +9,7 @@ if (class_exists('Database')) {
 }
 
 define('PROTECTOR_PRECHECK_INCLUDED', 1);
-define('PROTECTOR_VERSION', file_get_contents(__DIR__ . '/version.txt'));
+define('PROTECTOR_VERSION', (float) file_get_contents(__DIR__ . '/version.txt'));
 
 // set $_SERVER['REQUEST_URI'] for IIS
 if (empty($_SERVER['REQUEST_URI'])) {         // Not defined by IIS
