+++ htdocs/xoops_lib/modules/protector/xoops_version.php
@@ -22,7 +22,7 @@ $constpref = '_MI_' . strtoupper($mydirname);
 
 $modversion['name']           = constant($constpref . '_NAME');
 $modversion['description']    = constant($constpref . '_DESC');
-$modversion['version']        = file_get_contents(__DIR__ . '/include/version.txt');
+$modversion['version']        = (float) file_get_contents(__DIR__ . '/include/version.txt');
 $modversion['credits']        = 'PEAK Corp.';
 $modversion['author']         = 'GIJ=CHECKMATE<br>PEAK Corp.(http://www.peak.ne.jp/)';
 $modversion['help']           = 'page=help';
